move generic start/stopRecipe actions to containers thunks
